+++ src/main/java/io/github/astrapi69/swing/panels/lottery/LotteryPanel.java
@@ -33,6 +33,7 @@ import java.util.Map;
 
 import javax.swing.*;
 
+import io.github.astrapi69.icon.StringIcon;
 import io.github.astrapi69.model.BaseModel;
 import io.github.astrapi69.model.api.Model;
 import io.github.astrapi69.swing.base.BasePanel;