+++ src/test/java/io/github/astrapi69/swing/base/ApplicationPanelFrameExample.java
@@ -29,9 +29,9 @@ import java.awt.*;
 import javax.swing.*;
 import javax.swing.border.Border;
 
+import io.github.astrapi69.swing.button.IconButtonFactory;
 import org.jdesktop.swingx.MultiSplitLayout;
 
-import io.github.astrapi69.swing.button.IconButtonFactory;
 import io.github.astrapi69.swing.panels.splitpane.ApplicationTestModel;
 import io.github.astrapi69.swing.panels.splitpane.JXMultiSplitPanePanel;
 import io.github.astrapi69.swing.panels.splitpane.SplitFactory;